requirements.txt
streamlit
fastapi
python-dotenv
yt-dlp
requests
jinja2
python-multipart
nest-asyncio
uvicorn
starlette
aiofiles